Update patch set 1
Patch Set 1:

for the most part it is working.  I'm out of the country for about another month and haven't had a chance to test 14.1 on sprint but globally it was working.  I'm having an issue at the moment but I don't think it is related to the ril

Patch-set: 1
Reviewer: Kelly Craft <1690@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0