Update patch set 3
Patch Set 3:

> @Seth,
 > 
 > We might need to revert this if https://review.lineageos.org/#/c/64932/
 > doesn't get merged soon.  I would rather have 14.1 available even
 > without source built ril than keep it on 13.

Even with stock ril the issue with GsmCdmaPhone exists. If we can't get the other patch through, we need to investigate the carrierconfig in the phone app to see if we can determine why there is a second subid for gsm where the configuration is not loaded. That would also solve this issue.

Patch-set: 3